Rename constants to fixtures/cryptosuites.js.
